REFACTOR Query interface from Sfneal\Queries\Interfaces to Sfneal\Queries namesapce
