+++ src/Queries/NextOrPreviousModelQuery.php
@@ -5,7 +5,6 @@ namespace Sfneal\Queries;
 use Illuminate\Database\Eloquent\Model;
 use Sfneal\Caching\Traits\Cacheable;
 use Sfneal\Models\AbstractModel;
-use Sfneal\Queries\Interfaces\Query;
 
 class NextOrPreviousModelQuery implements Query
 {
